requirements.txt
alabaster==0.7.9
args==0.1.0
Babel==2.3.4
cffi==1.5.2
clint==0.5.1
colorama==0.3.7
cycler==0.9.0
decorator==4.0.6
docutils==0.12
docxtpl==0.3.5
enum34==1.1.2
future==0.15.2
idna==2.1
imageio==1.5
imagesize==0.7.1
ipaddress==1.0.16
Jinja2==2.8
lxml==3.6.1
MarkupSafe==0.23
matplotlib==1.5.1
moviepy==0.2.2.11
netaddr==0.7.18
numpy==1.10.4
palettable==2.1.1
pandas==0.17.1
pefile==2016.3.28
pelm==0.50
Pillow==3.1.0
pkginfo==1.3.2
pockets==0.3
py==1.4.31
pyasn1==0.1.9
pycparser==2.14
Pygments==2.1.3
PyInstaller==3.2
pyminifier==2.1
pyparsing==2.0.3
pypiwin32==219
pytest==2.8.5
python-dateutil==2.4.2
python-docx==0.8.6
pytz==2016.6.1
reportlab==3.2.0
requests==2.11.1
requests-toolbelt==0.7.0
scipy==0.16.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.5
sphinx-rtd-theme==0.1.9
sphinxcontrib-napoleon==0.5.2
tqdm==4.7.6
twine==1.8.1
